01-login/requirements.txt
flask>=2.0.3
authlib>=1.0
python-dotenv>=0.19.2
requests>=2.27.1
gunicorn>=20.1.0
